+++ example/spec/parallel/ingestGranule/IngestUMMGSuccessSpec.js
@@ -8,11 +8,8 @@ const {
 } = require('url');
 const mime = require('mime-types');
 
-const {
-  models: {
-    AccessToken, Execution, Provider
-  }
-} = require('@cumulus/api');
+const Execution = require('@cumulus/api/models/executions');
+const Provider = require('@cumulus/api/models/providers');
 const {
   getS3Object,
   s3ObjectExists,
@@ -21,23 +18,23 @@ const {
 } = require('@cumulus/aws-client/S3');
 const { generateChecksumFromStream } = require('@cumulus/checksum');
 const { constructCollectionId } = require('@cumulus/common/collection-config-store');
-const { LambdaStep } = require('@cumulus/integration-tests/sfnStep');
 const { getUrl } = require('@cumulus/cmrjs');
 const {
   addCollections,
-  api: apiTestUtils,
   buildAndExecuteWorkflow,
   conceptExists,
-  getOnlineResources,
-  granulesApi: granulesApiTestUtils,
-  EarthdataLogin: { getEarthdataAccessToken },
-  distributionApi: {
-    getDistributionApiRedirect,
-    getDistributionApiFileStream,
-    getDistributionFileUrl
-  }
+  getOnlineResources
 } = require('@cumulus/integration-tests');
+const apiTestUtils = require('@cumulus/integration-tests/api/api');
 const { deleteCollection } = require('@cumulus/integration-tests/api/collections');
+const granulesApiTestUtils = require('@cumulus/integration-tests/api/granules');
+const {
+  getDistributionFileUrl,
+  getTEADistributionApiRedirect,
+  getTEADistributionApiFileStream,
+  getTEARequestHeaders
+} = require('@cumulus/integration-tests/api/distribution');
+const { LambdaStep } = require('@cumulus/integration-tests/sfnStep');
 
 const {
   loadConfig,
@@ -93,7 +90,6 @@ describe('The S3 Ingest Granules workflow configured to ingest UMM-G', () => {
   let granule;
   let executionModel;
   let config;
-  let accessTokensModel;
   let testDataFolder;
   let collection;
   let provider;
@@ -110,8 +106,6 @@ describe('The S3 Ingest Granules workflow configured to ingest UMM-G', () => {
     provider = { id: `s3_provider${testSuffix}` };
     const newCollectionId = constructCollectionId(collection.name, collection.version);
 
-    process.env.AccessTokensTable = `${config.stackName}-AccessTokensTable`;
-    accessTokensModel = new AccessToken();
     process.env.GranulesTable = `${config.stackName}-GranulesTable`;
     process.env.ExecutionsTable = `${config.stackName}-ExecutionsTable`;
     executionModel = new Execution();
@@ -273,8 +267,8 @@ describe('The S3 Ingest Granules workflow configured to ingest UMM-G', () => {
     let onlineResources;
     let files;
     let resourceURLs;
-    let accessToken;
     let beforeAllError;
+    let teaRequestHeaders;
 
     beforeAll(async () => {
       postToCmrOutput = await lambdaStep.getStepOutput(workflowExecution.executionArn, 'PostToCmr');
@@ -289,18 +283,13 @@ describe('The S3 Ingest Granules workflow configured to ingest UMM-G', () => {
         process.env.CMR_ENVIRONMENT = 'UAT';
         const result = await Promise.all([
           getOnlineResources(granule),
-          // Login with Earthdata and get access token.
-          getEarthdataAccessToken({
-            redirectUri: process.env.DISTRIBUTION_REDIRECT_ENDPOINT,
-            requestOrigin: process.env.DISTRIBUTION_ENDPOINT
-          })
+          getTEARequestHeaders(config.stackName)
         ]);
 
         onlineResources = result[0];
         resourceURLs = onlineResources.map((resource) => resource.URL);
 
-        const accessTokenResponse = result[1];
-        accessToken = accessTokenResponse.accessToken;
+        teaRequestHeaders = result[1];
       } catch (e) {
         beforeAllError = e;
       }
@@ -308,10 +297,6 @@ describe('The S3 Ingest Granules workflow configured to ingest UMM-G', () => {
 
     beforeEach(() => {
       if (beforeAllError) fail(beforeAllError);
-    });
-
-    afterAll(async () => {
-      await accessTokensModel.delete({ accessToken });
     });
 
     it('has expected payload', () => {
@@ -375,16 +360,14 @@ describe('The S3 Ingest Granules workflow configured to ingest UMM-G', () => {
       expect(resourceURLs.includes(cumulusDocUrl)).toBe(true);
     });
 
-    // TODO Update this to use TEA
-    xit('includes the Earthdata login ID for requests to protected science files', async () => {
+    it('includes the Earthdata login ID for requests to protected science files', async () => {
       const filepath = `/${files[0].bucket}/${files[0].filepath}`;
-      const s3SignedUrl = await getDistributionApiRedirect(filepath, accessToken);
-      const earthdataLoginParam = new URL(s3SignedUrl).searchParams.get('x-EarthdataLoginUsername');
+      const s3SignedUrl = await getTEADistributionApiRedirect(filepath, teaRequestHeaders);
+      const earthdataLoginParam = new URL(s3SignedUrl).searchParams.get('A-userid');
       expect(earthdataLoginParam).toEqual(process.env.EARTHDATA_USERNAME);
     });
 
-    // TODO Update this to use TEA
-    xit('downloads the requested science file for authorized requests', async () => {
+    it('downloads the requested science file for authorized requests', async () => {
       const scienceFileUrls = resourceURLs.filter(isUMMGScienceUrl);
       console.log('scienceFileUrls: ', scienceFileUrls);
 
@@ -400,7 +383,7 @@ describe('The S3 Ingest Granules workflow configured to ingest UMM-G', () => {
             const file = files.find((f) => f.name.endsWith(extension));
 
             const filepath = `/${file.bucket}/${file.filepath}`;
-            const fileStream = await getDistributionApiFileStream(filepath, accessToken);
+            const fileStream = await getTEADistributionApiFileStream(filepath, teaRequestHeaders);
 
             // Compare checksum of downloaded file with expected checksum.
             const downloadChecksum = await generateChecksumFromStream('cksum', fileStream);