+++ CHANGELOG.md
@@ -123,7 +123,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
     be assigned to the Elasticsearch Domain.
 
 - **CUMULUS-1752**
-  - Added `@cumulus/integration-tests/distribution/getTEARequestHeaders` to generate necessary request headers for a request to the [Thin Egress App](https://github.com/asfadmin/thin-egress-app)
+  - Added `@cumulus/integration-tests/api/distribution.invokeTEADistributionLambda` to simulate a request to the [Thin Egress App](https://github.com/asfadmin/thin-egress-app) by invoking the Lambda and getting a response payload.
+  - Added `@cumulus/integration-tests/api/distribution.getTEARequestHeaders` to generate necessary request headers for a request to the Thin Egress App
+  - Added `@cumulus/integration-tests/api/distribution.getTEADistributionApiFileStream` to get a response stream for a file served by Thin Egress App
+  - Added `@cumulus/integration-tests/api/distribution.getTEADistributionApiRedirect` to get a redirect response from the Thin Egress App
 
 - **CUMULUS-1755**
   - Added `@cumulus/aws-client/CloudFormation.describeCfStack()` to describe a Cloudformation stack
@@ -194,7 +197,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 - **CUMULUS-1755**
   - Updated the Thin Egress App module used in `tf-modules/distribution/main.tf` to build 61. [See the release notes](https://github.com/asfadmin/thin-egress-app/releases/tag/tea-build.61).
-  - Updated `@cumulus/integration-tests/api/distribution.invokeApiDistributionLambda()` to invoke the Thin Egress App Lambda for testing
 
 - **CUMULUS-1757**
   - Update @cumulus/cmr-client CMRSearchConceptQueue to take optional cmrEnvironment parameter
@@ -219,6 +221,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 - **CUMULUS-1748**
   - Deprecate `@cumulus/ingest/util.normalizeProviderPath`
+
+- **CUMULUS-1752**
+  - Deprecate `@cumulus/integration-tests/api/distribution.getDistributionApiFileStream`
+  - Deprecate `@cumulus/integration-tests/api/distribution.getDistributionApiRedirect`
+  - Deprecate `@cumulus/integration-tests/api/distribution.invokeApiDistributionLambda`
 
 ### Removed